Make extract_current_risk_factors a shared_task
